+++ packages/api-client/index.ts
@@ -2,6 +2,7 @@ export { createApiClient } from "./create-api-client";
 export { createAdminDogsApiClient } from "./admin/dogs/create-admin-dogs-api-client";
 export { createAdminShowsApiClient } from "./admin/shows/create-admin-shows-api-client";
 export { createAdminTrialsApiClient } from "./admin/trials/create-admin-trials-api-client";
+export { createPublicCompetitionsApiClient } from "./public/competitions";
 export {
   searchPublicVirtualPairing,
   calculatePublicVirtualPairing,