+++ packages/server/competitions/index.ts
@@ -1,0 +1,3 @@
+export { createCompetitionsService, competitionsService } from "./service";
+export { getBestDriverRankingService } from "./best-driver";
+export type { CompetitionsServiceLogContext } from "./types";